chore: fix up tests for refactor
